refactor(webapp/auction): simplify Noun birthday and auction winner
