+++ packages/nouns-webapp/src/components/NounInfoRowBirthday.tsx
@@ -2,17 +2,15 @@ import React from 'react';
 
 import { i18n } from '@lingui/core';
 import { Trans } from '@lingui/react/macro';
-import { Image } from 'react-bootstrap';
 
-import _BirthdayIcon from '@/assets/icons/Birthday.svg';
 import { useAppSelector } from '@/hooks';
+import { cn } from '@/lib/utils';
 import { AuctionState } from '@/state/slices/auction';
 import { isNounderNoun } from '@/utils/nounderNoun';
 
-import classes from './NounInfoRowBirthday.module.css';
-
 interface NounInfoRowBirthdayProps {
   nounId: bigint;
+  className?: string;
 }
 
 export const getNounBirthday = (nounId: bigint, pastAuctions: AuctionState[]) => {
@@ -24,7 +22,7 @@ export const getNounBirthday = (nounId: bigint, pastAuctions: AuctionState[]) =>
   );
 };
 
-const NounInfoRowBirthday: React.FC<NounInfoRowBirthdayProps> = ({ nounId }) => {
+const NounInfoRowBirthday: React.FC<NounInfoRowBirthdayProps> = ({ nounId, className }) => {
   // If the noun is a nounder noun, use the next noun to get the mint date.
   // We do this because we use the auction start time to get the mint date and
   // nounder nouns do not have an auction start time.
@@ -43,15 +41,9 @@ const NounInfoRowBirthday: React.FC<NounInfoRowBirthdayProps> = ({ nounId }) => 
   const birthday = new Date(Number(startTime) * 1000);
 
   return (
-    <div className={classes.birthdayInfoContainer}>
-      <span>
-        <Image src={_BirthdayIcon} className={classes.birthdayIcon} />
-      </span>
-      <Trans>Born</Trans>
-      <span className={classes.nounInfoRowBirthday}>
-        {i18n.date(birthday, { month: 'long', year: 'numeric', day: '2-digit' })}
-      </span>
-    </div>
+    <span className={cn('text-muted-foreground block', className)}>
+      <Trans>Born {i18n.date(birthday, { month: 'long', year: 'numeric', day: '2-digit' })}</Trans>
+    </span>
   );
 };
